feat: add functionality to load whiteboard snapshot and thumbnail from database
